Reorganize files to be in line with bullet proof react pattern
