+++ src/app/pr-review/page.tsx
@@ -1,4 +1,4 @@
-import PullRequestViewer from "./PullRequestViewer";
+import PullRequestViewer from "../../features/pull-request/components/pr-viewer";
 
 export default function Page() {
   return <PullRequestViewer />;
